+++ source/Public/Compare-ResourcePropertyState.ps1
@@ -1,0 +1,154 @@
+<#
+    .SYNOPSIS
+        Compare current and desired property values for any DSC resource.
+
+    .DESCRIPTION
+        This function is used to compare current and desired property values for any
+        DSC resource, and return a hashtable with the metadata from the comparison.
+
+    .PARAMETER CurrentValues
+        The current values that should be compared to to desired values. Normally
+        the values returned from Get-TargetResource.
+
+    .PARAMETER DesiredValues
+        The values set in the configuration and is provided in the call to the
+        functions *-TargetResource, and that will be compared against current
+        values. Normally set to $PSBoundParameters.
+
+    .PARAMETER Properties
+        An array of property names, from the keys provided in DesiredValues, that
+        will be compared. If this parameter is left out, all the keys in the
+        DesiredValues will be compared.
+
+    .PARAMETER IgnoreProperties
+        An array of property names, from the keys provided in DesiredValues, that
+        will be ignored in the comparison. If this parameter is left out, all the
+        keys in the DesiredValues will be compared.
+
+    .PARAMETER CimInstanceKeyProperties
+        A hashtable containing a key for each property that contain a collection
+        of CimInstances and the value is an array of strings of the CimInstance
+        key properties.
+        @{
+            Permission = @('State')
+        }
+
+    .EXAMPLE
+        $compareTargetResourceStateParameters = @{
+            CurrentValues = (Get-TargetResource $PSBoundParameters)
+            DesiredValues = $PSBoundParameters
+        }
+
+        $propertyState = Compare-ResourcePropertyState @compareTargetResourceStateParameters
+
+        This examples call Compare-ResourcePropertyState with the current state
+        and the desired state and returns a hashtable array of all the properties
+        that was evaluated based on the properties pass in the parameter DesiredValues.
+#>
+function Compare-ResourcePropertyState
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable[]])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [System.Collections.Hashtable]
+        $CurrentValues,
+
+        [Parameter(Mandatory = $true)]
+        [System.Collections.Hashtable]
+        $DesiredValues,
+
+        [Parameter()]
+        [System.String[]]
+        $Properties,
+
+        [Parameter()]
+        [System.String[]]
+        $IgnoreProperties,
+
+        [Parameter()]
+        [System.Collections.Hashtable]
+        $CimInstanceKeyProperties = @{}
+    )
+
+    if ($PSBoundParameters.ContainsKey('Properties'))
+    {
+        # Filter out the parameters (keys) not specified in Properties
+        $desiredValuesToRemove = $DesiredValues.Keys |
+            Where-Object -FilterScript {
+                $_ -notin $Properties
+            }
+
+        $desiredValuesToRemove |
+            ForEach-Object -Process {
+                $DesiredValues.Remove($_)
+            }
+    }
+    else
+    {
+        <#
+            Remove any common parameters that might be part of DesiredValues,
+            if it $PSBoundParameters was used to pass the desired values.
+        #>
+        $commonParametersToRemove = $DesiredValues.Keys |
+            Where-Object -FilterScript {
+                $_ -in [System.Management.Automation.PSCmdlet]::CommonParameters `
+                    -or $_ -in [System.Management.Automation.PSCmdlet]::OptionalCommonParameters
+            }
+
+        $commonParametersToRemove |
+            ForEach-Object -Process {
+                $DesiredValues.Remove($_)
+            }
+    }
+
+    # Remove any properties that should be ignored.
+    if ($PSBoundParameters.ContainsKey('IgnoreProperties'))
+    {
+        $IgnoreProperties |
+            ForEach-Object -Process {
+                if ($DesiredValues.ContainsKey($_))
+                {
+                    $DesiredValues.Remove($_)
+                }
+            }
+    }
+
+    $compareTargetResourceStateReturnValue = @()
+
+    foreach ($parameterName in $DesiredValues.Keys)
+    {
+        Write-Debug -Message ($script:localizedData.EvaluatePropertyState -f $parameterName)
+
+        $parameterState = @{
+            ParameterName = $parameterName
+            Expected      = $DesiredValues.$parameterName
+            Actual        = $CurrentValues.$parameterName
+        }
+
+        # Check if the parameter is in compliance.
+        $isPropertyInDesiredState = Test-DscPropertyState -Values @{
+            CurrentValue = $CurrentValues.$parameterName
+            DesiredValue = $DesiredValues.$parameterName
+            KeyProperties = $CimInstanceKeyProperties.$parameterName
+        }
+
+        if ($isPropertyInDesiredState)
+        {
+            Write-Verbose -Message ($script:localizedData.PropertyInDesiredState -f $parameterName)
+
+            $parameterState['InDesiredState'] = $true
+        }
+        else
+        {
+            Write-Verbose -Message ($script:localizedData.PropertyNotInDesiredState -f $parameterName)
+
+            $parameterState['InDesiredState'] = $false
+        }
+
+        $compareTargetResourceStateReturnValue += $parameterState
+    }
+
+    return $compareTargetResourceStateReturnValue
+}