+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Added cmdlet `Compare-ResourcePropertyState` and also introduces a new
+  design pattern how to evaluate properties in both _Test_ and _Set_.
+
 ## [0.9.1] - 2020-07-08
 
 ## Added